airflow/requirements.txt
apache-airflow==2.8.1
networkx==2.8.8
pandas==1.5.3
psycopg2-binary==2.9.10
pymongo==4.9.1
requests==2.32.4
scikit-learn==1.3.2
